refactor: restructure pages schemas
